# docs/openapi/meal-plan-api.yaml
openapi: 3.0.3
info:
  title: Meal Plan Management Service API
  description: API for managing meal plans, recipes, and nutritional tracking
  version: 1.0.0
  contact:
    name: Recipe Web App Team
servers:
  - url: http://localhost:3000/api/v1/meal-plan-management
    description: Local development server
  - url: https://api.recipe-app.com/api/v1/meal-plan-management
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /meal-plans/tags:
    get:
      tags:
        - tags
      summary: List all tags
      description: Retrieve a paginated list of all available meal plan tags
      operationId: listMealPlanTags
      parameters:
        - $ref: '#/components/parameters/TagNameSearch'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of tags retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTagsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /meal-plans/{id}/tags:
    get:
      tags:
        - tags
      summary: Get meal plan tags
      description: Retrieve all tags associated with a specific meal plan
      operationId: getMealPlanTags
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      responses:
        '200':
          description: Tags retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlanTagsApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - tags
      summary: Add tags to meal plan
      description: >-
        Add tags to a meal plan. This is additive - existing tags are kept.
        New tags are created if they don't exist.
      operationId: addMealPlanTags
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMealPlanTagsDto'
      responses:
        '200':
          description: Tags added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlanTagsApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - tags
      summary: Replace meal plan tags
      description: Replace all tags on a meal plan. Existing tags are removed and replaced with the provided tags.
      operationId: replaceMealPlanTags
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMealPlanTagsDto'
      responses:
        '200':
          description: Tags replaced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealPlanTagsApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /meal-plans/{id}/tags/{tagId}:
    delete:
      tags:
        - tags
      summary: Remove tag from meal plan
      description: Remove a specific tag from a meal plan
      operationId: removeMealPlanTag
      parameters:
        - $ref: '#/components/parameters/MealPlanId'
        - $ref: '#/components/parameters/TagId'
      responses:
        '204':
          description: Tag removed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MealPlanId:
      name: id
      in: path
      required: true
      description: Meal plan ID
      schema:
        type: string
        example: '123'
...
    SortBy:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [name, startDate, endDate, createdAt, updatedAt]
        default: createdAt
    SortOrder:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: desc
...
    Page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TagId:
      name: tagId
      in: path
      required: true
      description: Tag ID
      schema:
        type: string
        example: '1'
    TagNameSearch:
      name: nameSearch
      in: query
      description: Search term for tag name
      schema:
        type: string
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: false
...
    MealPlanTagDto:
      type: object
      properties:
        tagId:
          type: string
          description: Tag ID
          example: '1'
        name:
          type: string
          maxLength: 50
          description: Tag name
          example: 'Weekly'
    AddMealPlanTagsDto:
      type: object
      required:
        - tags
      properties:
        tags:
          type: array
          items:
            type: string
            maxLength: 50
          minItems: 1
          maxItems: 20
          description: List of tag names to add
          example: ['Weekly', 'Budget']
    MealPlanTagsApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/MealPlanTagDto'
        message:
          type: string
          example: Tags updated successfully
    PaginatedTagsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/MealPlanTagDto'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Validation failed
            code:
              type: string
              example: VALIDATION_ERROR
            details:
              type: array
              items:
                type: string
              example:
                - Name must be at least 1 character long
                - End date must be after start date
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'